% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{build_json_list_string}
\alias{build_json_list_string}
\title{Title}
\usage{
build_json_list_string(fieldsvector)
}
\arguments{
\item{fieldsvector}{}
}
\description{
Title
}
\examples{
build_json_list_string(c("BID","ASK","OPEN_PRC"))
}
