% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{CheckandReportEmptyDF}
\alias{CheckandReportEmptyDF}
\title{function to check if a downloaded dataframe is empty}
\usage{
CheckandReportEmptyDF(df, functionname)
}
\arguments{
\item{df}{data.frame}

\item{functionname}{functionname for errorreporting}
}
\value{
boolean
}
\description{
function to check if a downloaded dataframe is empty
}
\examples{
CheckandReportEmptyDF(data.frame(), functionname = "test")
CheckandReportEmptyDF(data.frame(test = c(1,2),test2 = c("a","b")), functionname = "test")
}
