% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_custominstruments.r
\name{CheckifCustomInstrument}
\alias{CheckifCustomInstrument}
\title{Check if a symbol is really a custom symbol}
\usage{
CheckifCustomInstrument(symbol, UUID = getOption(".RefinitivUUID"))
}
\arguments{
\item{symbol}{character}

\item{UUID}{character  Eikon UUID can be found here: Eikon Terminal --> help --> about --> user details: UUID e.g. ABCDE-123456}
}
\value{
boolean TRUE or FALSE
}
\description{
Check if a symbol is really a custom symbol
}
\examples{
CheckifCustomInstrument(symbol = "test" , UUID = "ABCDE-123456")
CheckifCustomInstrument(symbol = "test.ABCDE-123456" , UUID = "ABCDE-123456")
}
\keyword{internal}
