% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{CheckINTERDAYEikonTimeSeries}
\alias{CheckINTERDAYEikonTimeSeries}
\title{A method to check if a downloaded INTERDAY timeseries if it returns the correct amount of dates.}
\usage{
CheckINTERDAYEikonTimeSeries(EikonTimeseriesResult, EikonDataResult)
}
\arguments{
\item{EikonTimeseriesResult}{a data.frame with timeseries generated with EikonGetTimeseries containing at least the columns "Date" and "Security"}

\item{EikonDataResult}{a data.frame with the columns "Instrument" denoting the RIC of the instrumennt and the column "Operating MIC" denoting the operating mic of the exchange on which the instrument is traded}
}
\description{
It checks if all dates that should be in the timeseries are also present.
The function automatically removes weekenddays and public Holidays.
The public Holidays that are taken into consideration are shown in Refinitiv::NagerHolidayData
}
\examples{
\dontrun{
Eikon <- Refinitiv::EikonConnect()
EikonDataWithFailingOPeratingMics <- EikonGetData(EikonObject = Eikon, rics = c("AAPL.O, GOOGL.O"),
                                           Eikonformulas = c("TR.OperatingMIC", "TR.CompanyName"))
EikonDataWithRepairedOPeratingMics <- EikonRepairMic(EikonDataWithFailingOPeratingMics)
EikonTimeseries <- EikonGetTimeseries(EikonObject = Eikon, rics = Stoxx1800Constits$RIC)
MissingDates <- CheckEikonTimeSeries(EikonTimeseriesResult = EikonTimeseries,
                                     EikonDataResult = EikonDataWithRepairedOPeratingMics )
}
}
