% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.r
\name{ConstructTicketJsonBody}
\alias{ConstructTicketJsonBody}
\title{rewrite JSON body in case a waiting ticket is assigned so that the correct json is requested}
\usage{
ConstructTicketJsonBody(query, ticket, debug)
}
\arguments{
\item{query}{httr2_response}

\item{ticket}{character hash code}

\item{debug}{boolean print revised json message with ticket number}
}
\value{
revised json body with ticket hash
}
\description{
rewrite JSON body in case a waiting ticket is assigned so that the correct json is requested
}
\examples{
\dontrun{
 see tests for examples
}
}
\keyword{internal}
