% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{create_OMM_Stream}
\alias{create_OMM_Stream}
\title{Create an OMM stream when an Eikon terminal is running}
\usage{
create_OMM_Stream(
  name = "EUR=",
  domain = "MarketPrice",
  service = NULL,
  fields = c("BID", "ASK", "OPEN_PRC"),
  extended_params = NULL
)
}
\arguments{
\item{name}{character  RIC to retrieve item stream.}

\item{domain}{character Specify item stream domain (MarketPrice, MarketByPrice, ...), defaults to MarketPrice}

\item{service}{Specify the service to subscribe on. NULL}

\item{fields}{character vector Specify the fields to retrieve.}

\item{extended_params}{NULL Specify optional params}
}
\value{
OMM stream object
}
\description{
Create an OMM stream when an Eikon terminal is running
}
\details{
Note that field names, depend on exchange and on order book label

\itemize{
 \item(domain = MarketByPrice [level2] )(ORDER_PRC, ORDER_SIDE, ACC_SIZE)
 \item(domain = MarketPrice)(BEST_BID1, BEST_BSIZ1, BEST_ASK1, BEST_ASIZ1)
}
}
\examples{
 \dontrun{
OMM_ws <- create_OMM_Stream()
EUR_stream <- OMM_ws$new(name = "EUR=", fields = c("BID","ASK","OPEN_PRC"))
# start stream
EUR_stream$connect()
# stop stream
EUR_stream$close()
}
}
