% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{create_RDP_Stream}
\alias{create_RDP_Stream}
\title{Create a RDP stream when an Eikon terminal is running}
\usage{
create_RDP_Stream(
  service,
  universe,
  view,
  parameters,
  api,
  extended_parameters
)
}
\value{
RDP Stream object
}
\description{
Create a RDP stream when an Eikon terminal is running
}
\examples{
\dontrun{
test <- create_RDP_Stream()
test$connect()
test$close()
}
}
