% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_custominstruments.r
\name{CustomInstrumentBasketBuilder}
\alias{CustomInstrumentBasketBuilder}
\title{Build a basket element for a custom instrument}
\usage{
CustomInstrumentBasketBuilder(RICs = NULL, Weights = NULL, Normalize = TRUE)
}
\arguments{
\item{Weights}{numeric vector of weights in the basket}

\item{Normalize}{boolean Normalize all weight to a maximum of 1?}

\item{RICS}{character vector of RICs in the basket}
}
\value{
a basket object that can be used
}
\description{
Build a basket element for a custom instrument
}
\examples{
CustomInstrumentBasketBuilder(RICs = c("AAPL.O", "AMZN.O"), Weights = c(0.5, 0.5))
}
\seealso{
[rd_ManageCustomInstruments()]
}
