% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_custominstruments.r
\name{CustomInstrumentHolidayBuilder}
\alias{CustomInstrumentHolidayBuilder}
\title{Build a holiday object for a custom instrument}
\usage{
CustomInstrumentHolidayBuilder(dates = NULL, reasons = NULL)
}
\arguments{
\item{dates}{vector of character dates in format YYYY-MM-DD}

\item{reasons}{character vector of reaons for this holida e.g. Christmas}
}
\value{
object with holidays
}
\description{
Build a holiday object for a custom instrument
}
\examples{
CustomInstrumentHolidayBuilder(dates = c("2023-12-01", "2023-12-31")
, reasons = c("Special Bank Holiday 1", "Special Bank Holiday 2"))
}
\seealso{
[rd_ManageCustomInstruments()]
}
