% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{DataStreamConnect}
\alias{DataStreamConnect}
\title{Initialize DataStream Connection}
\usage{
DataStreamConnect(DatastreamUserName, DatastreamPassword)
}
\arguments{
\item{DatastreamUserName}{Refinitiv DataStream username}

\item{DatastreamPassword}{Refinitiv DataStream password}
}
\description{
Initialize DataStream Connection
}
\examples{
\dontrun{
DatastreamUserName <- "Your datastream username"
DatastreamPassword <- "Your datastream password"
DataStream <- DataStreamConnect(DatastreamUserName, DatastreamPassword)
}
}
