% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{EikonErrorProcessor}
\alias{EikonErrorProcessor}
\title{ErrorProcessor for get_data python output}
\usage{
EikonErrorProcessor(Eikon_get_data_Error, Chunked, ChunkRowStart)
}
\arguments{
\item{Eikon_get_data_Error}{List intermediate output of EikonPostProcessor}

\item{Chunked}{Boolean indicating if the get_data python output is chunked or not}

\item{ChunkRowStart}{Vector what the cumulative position is in the final data output}
}
\value{

}
\description{
This function processes the error output of the python function in r readable output.
This function is used by
}
\seealso{
\link{EikonGetData},
         \link{EikonGetTimeseries}
}
