% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{EikonShowAttributes}
\alias{EikonShowAttributes}
\title{Show the attributes of the Python Eikon}
\usage{
EikonShowAttributes(EikonObject)
}
\arguments{
\item{EikonObject}{Python Object generated by EikonConnect}
}
\description{
Function that the returns the names of the python Eikon api attributes that can be used as commands in R.
}
\examples{
\dontrun{
Eikon <- EikonConnect(Eikonapplication_id = "your key", Eikonapplication_port = 9000L)
EikonShowAttributes(EikonObject = Eikon)
}
}
