% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{EikonTimeSeriesPreprocessor}
\alias{EikonTimeSeriesPreprocessor}
\title{Preprocessor for Eikon Get timeseries to automaticallt chunk pieces in the required length}
\usage{
EikonTimeSeriesPreprocessor(interval, rics, start_date, end_date)
}
\arguments{
\item{interval}{character choose from  c('tick', 'minute', 'hour', 'daily', 'weekly', 'monthly', 'quarterly', 'yearly')}

\item{rics}{character vector containing the Reuters unique stock indentifier}

\item{start_date}{character start date in format "YYYY-MM-DD"}

\item{end_date}{character end date in format "YYYY-MM-DD"}
}
\value{

}
\description{
Preprocessor for Eikon Get timeseries to automaticallt chunk pieces in the required length
}
\examples{
test <- EikonTimeSeriesPreprocessor(interval = "daily", rics = rep(letters, 1000), start_date = "2015-01-01", end_date = "2018-01-01")
}
