% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{flattenNestedlist}
\alias{flattenNestedlist}
\title{Flatten a nested list put list format in data.table format way (don't mix up lists and vectors in one nested list)}
\usage{
flattenNestedlist(data)
}
\arguments{
\item{data}{a nested list}
}
\value{
a list
}
\description{
Flatten a nested list put list format in data.table format way (don't mix up lists and vectors in one nested list)
}
\examples{
# Refinitiv:::flattenNestedlist(list(list("a", "b"), c(1,2)))
}
\keyword{internal}
