% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{get_rdp_streaming_url}
\alias{get_rdp_streaming_url}
\title{Show all available custom instruments that have been created}
\usage{
get_rdp_streaming_url(RDObject = RefinitivJsonConnect(), debug = TRUE)
}
\arguments{
\item{RDObject}{Refinitiv Data connection object, defaults to RefinitivJsonConnect()}

\item{debug}{show api calls defaults to FALSE}
}
\value{
a list of custom Instruments created with all parameters
}
\description{
Show all available custom instruments that have been created
}
\examples{
test <- get_rdp_streaming_url()
}
