% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{ImportCustomPythonutils}
\alias{ImportCustomPythonutils}
\title{Import The package Custom python utils for handling pre processing python rdp outputs}
\usage{
ImportCustomPythonutils()
}
\value{
python module
}
\description{
Import The package Custom python utils for handling pre processing python rdp outputs
}
\examples{
utils <- Refinitiv:::ImportCustomPythonutils()
}
\keyword{internal}
