% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{indexExists}
\alias{indexExists}
\title{Check if an element of list exists or is NULL}
\usage{
indexExists(list, index)
}
\arguments{
\item{list}{a list}

\item{index}{integer, indicating proposed list location}
}
\value{
boolean
}
\description{
Check if an element of list exists or is NULL
}
\examples{
Testlist <- list(NULL, 1)
indexExists(Testlist, 1)
}
