% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{InspectRequest}
\alias{InspectRequest}
\title{function to check if a downloaded dataframe is empty}
\usage{
InspectRequest(df, functionname, verbose = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{functionname}{function name for error reporting}

\item{verbose}{Boolean print or not variable}
}
\value{
boolean
}
\description{
function to check if a downloaded dataframe is empty
}
\examples{
InspectRequest(data.frame(), functionname = "test")
InspectRequest(data.frame(test = c(1,2),test2 = c("a","b")), functionname = "test")
}
