% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeCleaner.r
\name{make.true.NA_df}
\alias{make.true.NA_df}
\title{Remove NULL items, <NA>, NaN, and "" from dataframes and replace with NA while retaining column class}
\usage{
make.true.NA_df(df)
}
\arguments{
\item{df}{data.frame with/without  <NA> NaN ""}
}
\value{
data.frame with only NA entries in the right column class
}
\description{
the function only operates on character and numeric columns
}
\examples{
dfr<-data.frame(A=c(1,2,NaN,3),B=c("a", "<NA>" ,"","d")
, C = c(as.Date("2012-01-01"), as.Date("2012-01-02"), as.Date("2012-01-03"), as.Date("2012-01-04")))
make.true.NA_df(dfr)
}
