% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapEikonTimefieldsToRd.R
\name{mapEikonTimefieldsToRd}
\alias{mapEikonTimefieldsToRd}
\title{Map Eikon Time Fields to Refinitiv Data (RD) Fields}
\usage{
mapEikonTimefieldsToRd(
  cols = c("TIMESTAMP", "VOLUME", "HIGH", "LOW", "OPEN", "CLOSE")
)
}
\arguments{
\item{cols}{A character vector of original column names in Eikon format. Defaults to c("TIMESTAMP", "VOLUME", "HIGH", "LOW", "OPEN", "CLOSE").}
}
\value{
A character vector of translated column names for Refinitiv Data (RD).
}
\description{
This function translates a vector of column names from the Eikon format to corresponding Refinitiv Data (RD) names based on a predefined mapping. If a column name does not have a mapping, it is kept unchanged.
}
\examples{
mapEikonTimefieldsToRd()
mapEikonTimefieldsToRd(c("TIMESTAMP", "UNKNOWN_COLUMN", "CLOSE"))

}
\seealso{
[rd_GetHistoricalPricing()]
}
