% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{NagerHolidayData}
\alias{NagerHolidayData}
\title{Public Holidays between the years 1990 and 2030}
\format{
A list containing:
\describe{
  \item{NagerHolidayData}{a data.frame with the specifics of the Holiday dates by country and year}
  \item{HolidayVectorperCountry}{a list of vectors containing the public holiday dates per country}
  \item{HolidayDataFrame}{a dataframe with two columns date (of the holiday) and the iso3c countryindication of the country where the holiday is located}
}
}
\source{
\url{https://date.nager.at/}
}
\usage{
NagerHolidayData
}
\description{
A dataset containing the public holidays for 100 Countries.
}
\keyword{datasets}
