% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{OperatingMicLookup}
\alias{OperatingMicLookup}
\title{Operating Mic lookup from Reuters Exchange Code}
\format{
A data.frame containing:
\describe{
  \item{RDN_EXCHD2}{Reuters Exchange Code}
  \item{repairedMIC}{An operating mic for this Reuters exchange Code}
}
}
\source{
derived from Reuters Eikon terminal
}
\usage{
OperatingMicLookup
}
\description{
A data.frame containing a lookup list from Reuters exchange codes to operating mics
}
\keyword{datasets}
