% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{PostProcessTimeSeriesRequest}
\alias{PostProcessTimeSeriesRequest}
\title{Postprocessor for raw Timeseries Requests}
\usage{
PostProcessTimeSeriesRequest(RawTimeSeriesRequest)
}
\arguments{
\item{RawTimeSeriesRequest}{raw return from eikon EikonGetTimeseries}
}
\value{
data.frame
}
\description{
Postprocessor for raw Timeseries Requests
}
\examples{
\dontrun{
RawTimeSeries <- try(EikonGetTimeseries( EikonObject = Eikon,
rics = c("MMM"),
start_date = "2020-01-01T01:00:00",
end_date = "2020-01-10T01:00:00",
fields = "CLOSE", raw = TRUE))1
PostProcessTimeSeriesRequest(RawTimeSeries)
}
}
\keyword{internal}
