% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EikonResult.R
\name{print.EikonResult}
\alias{print.EikonResult}
\title{Print an EikonResult}
\usage{
\method{print}{EikonResult}(x, ...)
}
\arguments{
\item{x}{An \code{EikonResult} object.}

\item{...}{Ignored.}
}
\value{
\code{x} (invisibly).
}
\description{
Print an EikonResult
}
