% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{Process_RDP_output}
\alias{Process_RDP_output}
\title{Output processor for python function RDP$get_search_metadata}
\usage{
Process_RDP_output(python_json, RemoveNA = FALSE, CleanNames = FALSE)
}
\arguments{
\item{python_json}{python json string}

\item{RemoveNA}{boolean remove NA value's defaults to FALSE}

\item{CleanNames}{boolean clean names of data.table, defaults to FALSE}
}
\value{
r data.frame
}
\description{
Output processor for python function RDP$get_search_metadata
}
\examples{
\dontrun{
"internal function don't use directly"
path = paste0("C:\\\\Users\\\\XXXX\\\\Documents\\\\GitHub",
"\\\\RefinitivR\\\\tests\\\\testthat\\\\PY_get_search_metadata_input.py")
PY_get_search_metadata_input <- reticulate::r_to_py(
 reticulate::py_load_object(file = path))
 r_df <- Process_RDP_output(PY_get_search_metadata_input)
}
}
\seealso{
[get_search_metadata()]

[rd_GetHistory()]
}
\keyword{internal}
