% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{ProcessIPAOutput}
\alias{ProcessIPAOutput}
\title{Process output of RDP IPA calls}
\usage{
ProcessIPAOutput(IPAoutput)
}
\arguments{
\item{IPAoutput}{output to a python call RDP$ipa$FinancialContracts}
}
\value{
data.frame
}
\description{
Process output of RDP IPA calls
}
\examples{
\dontrun{
ipa_output <- RDPGetOptionAnalytics(
 OptionRics = c("AAPLL032112500.U", "AAPLL032113700.U")
, raw = TRUE)
ProcessIPAOutput(ipa_output)
}
}
\keyword{internal}
