% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{ProcessSymbology}
\alias{ProcessSymbology}
\title{Function to process raw output of python get_symbology to better in r readable format}
\usage{
ProcessSymbology(EikonSymbologyResult, from_symbol_type, to_symbol_type)
}
\arguments{
\item{EikonSymbologyResult}{nested list: output from EikonGetSymbology with option raw_output set to TRUE}

\item{from_symbol_type}{character use her same input as in EikonGetSymbology}

\item{to_symbol_type}{character use her same input as in EikonGetSymbology}
}
\value{
data.frame containing 4 columns to_symbol_type, from_symbol_type, BestMatch (as defined by EIkon), error
}
\description{
Function to process raw output of python get_symbology to better in r readable format
}
\examples{
\dontrun{
Raw_output_No_BestMatch <- EikonGetSymbology(EikonObject = Eikon
, symbol =  c("GB00B03MLX29", "US0378331005"), from_symbol_type = "ISIN"
, to_symbol_type = "RIC" , raw_output = TRUE, bestMatch = FALSE  )
ProcessSymbology(EikonSymbologyResult = Raw_output_No_BestMatch
, from_symbol_type = "ISIN", to_symbol_type = "RIC")

Raw_output_BestMatch <- EikonGetSymbology(EikonObject = Eikon
, symbol =  c("GB00B03MLX29", "US0378331005"), from_symbol_type = "ISIN"
, to_symbol_type = "RIC" , raw_output = TRUE, bestMatch = TRUE  )
ProcessSymbology(EikonSymbologyResult = Raw_output_BestMatch
, from_symbol_type = "ISIN", to_symbol_type = "RIC")
}
}
