% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{PropertiesActiveRefinitivObject}
\alias{PropertiesActiveRefinitivObject}
\title{Function to check through which Refinitiv Connection object requests are made}
\usage{
PropertiesActiveRefinitivObject(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean defaults to TRUE}
}
\value{
named list
}
\description{
Function to check through which Refinitiv Connection object requests are made
}
\examples{
test <- PropertiesActiveRefinitivObject()
}
