% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{rd_check_proxy_url}
\alias{rd_check_proxy_url}
\title{Check if refinitiv proxy url is alive}
\usage{
rd_check_proxy_url(port = 9060, debug = TRUE)
}
\arguments{
\item{port}{9060 or 9000}

\item{debug}{boolean TRUE or False}
}
\value{
api status code and version
}
\description{
Check if refinitiv proxy url is alive
}
\examples{
\dontrun{
test <- rd_check_proxy_url(port = 9000)
test <- rd_check_proxy_url(port = 9060)
}
}
