% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{rd_handshake}
\alias{rd_handshake}
\title{Get bearer key from terminal (experimental, don't use in production)}
\usage{
rd_handshake(debug = FALSE)
}
\arguments{
\item{debug}{debug parameter print urls for json requests}
}
\value{
list with the following fields \itemize{
 \item(access_token)(key)
 \item(expires_in)(unknown what this means)
 \item(token_type)(type of token (bearer))
}
}
\description{
Get bearer key from terminal (experimental, don't use in production)
}
\examples{
\dontrun{
 rd_handshake()
}
}
