% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalPricing.r
\name{rd_OutputProcesser}
\alias{rd_OutputProcesser}
\title{Process output from refintiv data to r data.frame output}
\usage{
rd_OutputProcesser(x, use_field_names_in_headers = TRUE, NA_cleaning = TRUE)
}
\arguments{
\item{x}{refinitiv data platform output}

\item{use_field_names_in_headers}{boolean wheater or not to return titles of field (formulas) as headers}

\item{NA_cleaning}{clean NA in return data}
}
\value{
data.frame
}
\description{
Process output from refintiv data to r data.frame output
}
\examples{
\dontrun{
 EndPoint = "data/datagrid/beta1/"
 payload <- list( 'universe'= as.list(c("GOOG.O", "NVDA.O"))
               , 'fields'= as.list(c('TR.CLOSE', 'TR.OPEN'))
               , 'parameters'=list('SDate'= '2022-10-05', 'EDate'= '2022-11-05')
               , 'output'= 'Col,T|Va,Row,In,date|'
               )

response <- send_json_request(json = payload, service = "rdp"
, EndPoint = EndPoint, request_type = "POST")
Output <- rd_OutputProcesser(response)
}
}
\keyword{internal}
