% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_custominstruments.r
\name{rd_SearchCustomInstruments}
\alias{rd_SearchCustomInstruments}
\title{Show all available custom instruments that have been created}
\usage{
rd_SearchCustomInstruments(RDObject = RefinitivJsonConnect(), debug = TRUE)
}
\arguments{
\item{RDObject}{Refinitiv Data connection object, defaults to RefinitivJsonConnect()}

\item{debug}{show api calls defaults to false}
}
\value{
a list of custom Instruments created with all parameters
}
\description{
Show all available custom instruments that have been created
}
\examples{
\dontrun{
test <- rd_SearchCustomInstruments()
}
}
