% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming_debug.R
\name{rd_streaming_debug}
\alias{rd_streaming_debug}
\title{Debug streaming connection}
\usage{
rd_streaming_debug(stream, verbose = FALSE)
}
\arguments{
\item{stream}{Stream object}

\item{verbose}{Show detailed message information}
}
\value{
List with debug information
}
\description{
Helper function to debug streaming issues by showing what messages are received
}
