% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_streaming.r
\name{rdp_get_snapshot}
\alias{rdp_get_snapshot}
\title{Retrieve a real-time snapshot from a previously defined stream object}
\usage{
rdp_get_snapshot(StreamingPricesObject)
}
\arguments{
\item{StreamingPricesObject}{an object as defined by}
}
\value{
a data.table with snapshot prices
}
\description{
Retrieve a real-time snapshot from a previously defined stream object
}
\examples{
\dontrun{
rdp <- Refinitiv::RDPConnect("your key")
streamObject <- rdp_streaming_prices(rdp, universe = c("EUR=","JPY="), fields = c('DSPLY_NAME', 'BID', 'ASK', 'CF_TIME'))
rdp_get_snapshot(streamObject)
}
}
