% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_streaming.r
\name{rdp_streaming_prices}
\alias{rdp_streaming_prices}
\title{Define a streaming prices object for use in the get_snapshot function}
\usage{
rdp_streaming_prices(
  rdpObject,
  universe = list("EUR=", "JPY=", "GBP=", "HKD="),
  fields = list("DSPLY_NAME", "BID", "ASK", "CF_TIME")
)
}
\arguments{
\item{rdpObject}{refinitiv dataplatform object result from a call to Refinitiv::RDPConnect}

\item{universe}{list of instruments to be subscribed}

\item{fields}{list of fields to be returned by the stream}
}
\value{
a python stream connection object
}
\description{
Define a streaming prices object for use in the get_snapshot function
}
\examples{
\dontrun{
rdp <- Refinitiv::RDPConnect("your key")
streamObject <- rdp_streaming_prices(rdp, universe = c("EUR=","JPY="), fields = c('DSPLY_NAME', 'BID', 'ASK'))
EquityStream <- rdp_streaming_prices(rdp, universe = c("FLOW.AS"), fields = c('DSPLY_NAME', 'BID', 'ASK'))
}
}
\seealso{
rdp_get_snapshot
}
