% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{RDPConnect}
\alias{RDPConnect}
\title{RDP connection function to refinitiv Dataplatform libraries}
\usage{
RDPConnect(application_id = NA, PythonModule = "JSON")
}
\arguments{
\item{application_id}{refinitiv dataplatform api key}

\item{PythonModule}{character either RDP (python) or JSON (direct json message)}
}
\value{
rdp opbject
}
\description{
RDP connection function to refinitiv Dataplatform libraries
}
\examples{
\dontrun{
rdp <- RDPConnect(application_id = "your key")
}
}
