% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{RDPGetOptionAnalytics}
\alias{RDPGetOptionAnalytics}
\title{Get RDP option analytics}
\usage{
RDPGetOptionAnalytics(RDP = RDPConnect(), OptionRics = NULL, raw = FALSE)
}
\arguments{
\item{RDP}{Refinitiv DataPlatform Connection object, defaults to  RDPConnect()}

\item{OptionRics}{character vector with option rics}

\item{raw}{return raw data from RDP call}
}
\value{
data.frame with option data
}
\description{
Get RDP option analytics
}
\examples{
\dontrun{
RDPGetOptionAnalytics(OptionRics = c("AAPLL032112500.U", "AAPLL032113700.U"))
}
}
