% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{RDPsearch}
\alias{RDPsearch}
\title{RDP search function is a wrapper for the pyton rdp.search function}
\usage{
RDPsearch(
  RDP = RDConnect(),
  query = NULL,
  view = NULL,
  select = NULL,
  top = NULL,
  filter = NULL,
  boost = NULL,
  order_by = NULL,
  group_by = NULL,
  group_count = NULL,
  navigators = NULL,
  features = NULL,
  Arglist = list()
)
}
\arguments{
\item{RDP}{Refinitiv DataPlatform Connection object}

\item{query}{optional character}

\item{view}{optional character see also RDPShowAvailableSearchViews for available searchviews}

\item{select}{optional character string of length 1 e.g/ "ContractType,RIC"}

\item{top}{optional numeric search result cut off}

\item{filter}{optional character filter e.g. "startswith(LastName,'H')"}

\item{boost}{optional meaning not clear from refinitiv documentation}

\item{order_by}{optional character string e.g. 'DocumentTitle asc'}

\item{group_by}{optional character string e.g. 'FirstName'}

\item{group_count}{optional numeric number of items displayed per group}

\item{navigators}{optional character string e.g.}

\item{features}{optional character, meaning not clear from refinitiv documentation}

\item{Arglist}{optional named list pass the above parameters as a named list withouding needing to use to do.call.}
}
\value{
data.frame with search results
}
\description{
RDP search function is a wrapper for the pyton rdp.search function
}
\details{
For additional examples see  \url{https://github.com/Refinitiv-API-Samples/Article.RDPLibrary.Python.Search}
}
\examples{
\dontrun{
RDPConnect('your api key')
test <- RDPsearch(query =  "AAPL.O", select = "ContractType,RIC")

Presidents <- RDPsearch( view = "People", query = 'president'
                       , filter = "startswith(LastName,'H')"
                       , select = 'DocumentTitle'
                       , boost = ''
                       , order_by = 'DocumentTitle asc'
                       , group_by = 'FirstName'
                       , group_count = 2
                       , top = 20
                       , navigators = 'HullType'
                       , features = 'spell' )

reporates <- RDPsearch( view = "IndicatorQuotes"
                      , query = "repo rate", group_by = "CentralBankName"
                      , group_count = 3
                      , select = paste0("CentralBankName,DocumentTitle,"
                                       ,"RIC,ObservationValue")
                      , top = 1000)

EquitiesSearch <-  RDPsearch( view = "EquityQuotes"
                            , filter = paste0("Eps gt 6.0 and "
                                      , "RCSTRBC2012Name eq 'Personal & "
                                      , "Household Products & Services' "
                                      , "and MktCapTotalUsd gt 100000000 "
                                      , "and IsPrimaryRIC eq true")
                            , top =  10000
                            , select = paste0("DocumentTitle , RIC, Eps,"
                                             ," MktCapTotalUsd"))


Vessels <- RDPsearch( view = "VesselPhysicalAssets"
                    , filter = paste0( "RCSAssetTypeLeaf eq 'tanker'"
                               , " and RCSRegionLeaf eq 'Gulf of Mexico'")
                    , top =  10000
                    , navigators = "OriginPort"
                    , select = paste0( "DocumentTitle,RIC,OriginPort"
                                     , " ,DestinationPort,RCSFlagLeaf"
                                     , ",AssetName,AISStatus,"
                                     , "VesselCurrentPortRIC,IMO")
                    )


ListedSearch <- RDPsearch(Arglist = list(query = "president", view = "People"))

SearchQuery = "aapl.o"
ListedSearch <- RDPsearch(query = SearchQuery)
}

\dontrun{
  SearchQuery = "aapl.o"
  ListedSearch <- RDPsearch(RDP = RefinitivJsonConnect(), query = SearchQuery)

}

}
\seealso{
RDPShowAvailableSearchViews()
}
