% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_vault.R
\name{refinitiv_vault_has}
\alias{refinitiv_vault_has}
\title{Check whether a credential exists in the vault}
\usage{
refinitiv_vault_has(key)
}
\arguments{
\item{key}{Character scalar.}
}
\value{
Logical.
}
\description{
Check whether a credential exists in the vault
}
\keyword{internal}
