% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.r
\name{RefinitivJsonConnect}
\alias{RefinitivJsonConnect}
\title{Connect to Eikon directly with JSON requests}
\usage{
RefinitivJsonConnect(Eikonapplication_id = NA, Eikonapplication_port = 9000L)
}
\arguments{
\item{Eikonapplication_id}{character eikon api key}

\item{Eikonapplication_port}{numeric proxy port 9000 or 9060 for RDP/RD}
}
\value{
RefinitivConnection Object
}
\description{
Connect to Eikon directly with JSON requests
}
\examples{
\dontrun{
 EikonJson <- RefinitivJsonConnect()
}
}
