% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{replaceInList}
\alias{replaceInList}
\title{Replace items in nested list}
\usage{
replaceInList(x, FUN, ...)
}
\arguments{
\item{x}{list}

\item{FUN}{function to operate on list}

\item{...}{pass through parameters}
}
\value{
list
}
\description{
Replace items in nested list
}
\examples{
 x <- list(list(NA, NULL, NULL), list("a", "b", "c"))
test <- Refinitiv:::replaceInList(x, function(x)if(is.null(x))NA else x)
}
\keyword{internal}
