% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{retry}
\alias{retry}
\title{Function to retry failed functions after a time out of 5 seconds. Especially useful for failed api calls.}
\usage{
retry(retryfun, max = 10, init = 0)
}
\arguments{
\item{retryfun}{function to retry}

\item{max}{maximum number of retries, default = 0}

\item{init}{initial state of retries should always be zero}
}
\value{

}
\description{
Function to retry failed functions after a time out of 5 seconds. Especially useful for failed api calls.
}
\examples{
 retry(sum(1,"a"), max = 2)
}
