% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SearchViewsLookup}
\alias{SearchViewsLookup}
\title{Lookup code for SearchViews to alternate between RD and JSON format SearchViews}
\format{
A data.table containing:
\describe{
  \item{SearchViews_RD}{SearchViews valid for RD python API}
  \item{SearchViews_JSON_RDP}{SearchViews valid for RDP python API and direct JSON calls}
}
}
\source{
Refinitiv Api Documentation
}
\usage{
SearchViewsLookup
}
\description{
A data.table containing a lookup list from RD SearchViews to JSON SearchViews
}
\keyword{datasets}
