% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStreamConnection.r
\name{TestDataStreamCredentials}
\alias{TestDataStreamCredentials}
\title{Test if the DataStream Credentials are valid}
\usage{
TestDataStreamCredentials(DatastreamUsername = NULL, DatastreamPassword = NULL)
}
\arguments{
\item{DatastreamUsername}{Character. The DataStream username.}

\item{DatastreamPassword}{Character. The DataStream password.}
}
\value{
Logical. Returns `TRUE` if the credentials are valid. If invalid,
  returns `FALSE` and issues a warning with the error message.
}
\description{
This function checks the validity of the provided DataStream username and
password by sending a request to the DataStream API.
}
\examples{
TestDataStreamCredentials(DatastreamUsername = "wrongusername"
, DatastreamPassword = "wrongPassword")
}
