% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{TR_Field}
\alias{TR_Field}
\title{Helper function to build the Eikonformulas parameter for the EikonGetData function.}
\usage{
TR_Field(
  Field_name = NULL,
  Parameters = NULL,
  sort_dir = "asc",
  sort_priority = 0
)
}
\arguments{
\item{Field_name}{string Field name to request. You can find the list in Data Item Browser.}

\item{Parameters}{named List containing the parameters for the field passed in the argument field_name}

\item{sort_dir}{string Indicate the sort direction. Possible values are \'asc\' or \'desc\'. The default value is \'asc\'}

\item{sort_priority}{integer Gives a priority to the field for the sorting. The highest priority is 0 (zero). The default value is NULL}
}
\value{
a list of list which can be used as import for
}
\description{
Helper function to build the Eikonformulas parameter for the EikonGetData function.
}
\examples{
TR_Field(Field_name = 'tr.revenue')
TR_Field(Field_name ='tr.open', sort_dir ='asc', sort_priority = 1)
TR_Field(Field_name ='TR.GrossProfit', Parameters = list('Scale' = 6, 'Curn'= 'EUR')
        , sort_dir = 'asc', sort_priority = 0)
}
