% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{WeekendDates}
\alias{WeekendDates}
\title{Function to determine weekend Dates}
\usage{
WeekendDates(MinDate, MaxDate)
}
\arguments{
\item{MinDate}{Start date in character of Date class format = "\%Y-\%m-\%d"}

\item{MaxDate}{End date in character of Date class format = "\%Y-\%m-\%d"}
}
\value{
a vector of weekend dates between MinDate and MaxDate.
}
\description{
Function to determine weekend Dates
}
\examples{
WeekendDates(MinDate = "2013-01-01", MaxDate = "2019-01-01")
WeekendDates(MinDate = as.Date("2013-01-01"), MaxDate = as.Date("2019-01-01"))
}
