% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_streaming.r
\name{create_RDP_Stream}
\alias{create_RDP_Stream}
\title{Create a RDP stream when an Eikon terminal is running}
\usage{
create_RDP_Stream(
  service,
  universe,
  view,
  parameters,
  api,
  extended_parameters
)
}
\arguments{
\item{service}{string optional name of RDP service}

\item{universe}{RIC to retrieve item stream.}

\item{view}{data fields to retrieve item stream}

\item{parameters}{extra parameters to retrieve item stream.}

\item{api}{specific name of RDP streaming defined in config file. 'streaming/trading-analytics/redi'}

\item{extended_parameters}{defaults to NULL Specify optional params}
}
\value{
RDP Stream object
}
\description{
Create a RDP stream when an Eikon terminal is running
}
\examples{
\dontrun{
test <- create_RDP_Stream()
test$connect()
test$close()
}
}
