% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_custominstruments.r
\name{CustomInstrumentUDCBuilder}
\alias{CustomInstrumentUDCBuilder}
\title{Build a custom instrument using a user defined continuation}
\usage{
CustomInstrumentUDCBuilder(root, rollover, spreadAdjustment)
}
\arguments{
\item{root}{CC or CL}

\item{spreadAdjustment}{}
}
\description{
Build a custom instrument using a user defined continuation
}
\seealso{
[rd_ManageCustomInstruments()]
}
