% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{EikonConnect}
\alias{EikonConnect}
\title{Initialize Eikon Python api}
\usage{
EikonConnect(
  Eikonapplication_id = NA,
  Eikonapplication_port = 9000L,
  UUID = NA,
  PythonModule = "JSON",
  TestConnection = FALSE
)
}
\arguments{
\item{Eikonapplication_id}{eikon api key}

\item{Eikonapplication_port}{proxy port id}

\item{UUID}{optional character parameter for custom instruments, not necessairy for regular requests}

\item{PythonModule}{character choose between Eikon (python),RDP (python),JSON (direct JSON message without python)}

\item{TestConnection}{boolean, TRUE or FALSE test connection after initiating contact with Eikon terminal}
}
\value{
a Python module that is an EikonObject
}
\description{
Initialize Eikon Python api
}
\examples{
\dontrun{
Eikon <- EikonConnect(Eikonapplication_id = "your key", Eikonapplication_port = 9000L
, PythonModule = "Eikon")
Eikon <- EikonConnect(Eikonapplication_id = "your key", Eikonapplication_port = 9000L
, PythonModule = "RDP")
}
}
