% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{EikonConnect}
\alias{EikonConnect}
\title{Initialize Eikon Python api}
\usage{
EikonConnect(Eikonapplication_id = .EikonApiKey, Eikonapplication_port = 9000L)
}
\arguments{
\item{Eikonapplication_id}{eikon api key}

\item{Eikonapplication_port}{proxy port id}
}
\description{
Initialize Eikon Python api
}
\examples{
\dontrun{
Eikon <- EikonConnect(Eikonapplication_id = "your key", Eikonapplication_port = 9000L)
}
}
