% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{EikonGetData}
\alias{EikonGetData}
\title{Function to obtain data from Eikon. Based on the Eikon python function get_data}
\usage{
EikonGetData(EikonObject, rics, Eikonformulas, raw_output = FALSE)
}
\arguments{
\item{EikonObject}{Eikon object created using EikonConnect function}

\item{rics}{a vector containing the instrument RICS}

\item{Eikonformulas}{a vector containing character string of Eikon Formulas}

\item{raw_output}{to return the raw list by chunk for debugging purposes, default = FALSE}
}
\value{

}
\description{
The function automatically chunks the list of rics into chunks that comply with the api limitations and in the end rebuilds the chunks again into a single data.frame.
}
\examples{
\dontrun{
Eikon <- Refinitiv::EikonConnect()
EikonGetData(EikonObject = Eikon, rics = c("MMM", "III.L"),
             Eikonformulas = c("TR.PE(Sdate=0D)/*P/E (LTM) - Diluted Excl*/", "TR.CompanyName"))
}
}
\references{
\url{https://developers.refinitiv.com/eikon-apis/eikon-data-api/docs?content=49692&type=documentation_item}
}
