% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{EikonNameCleaner}
\alias{EikonNameCleaner}
\title{Convert eikon formula's in human readable names}
\usage{
EikonNameCleaner(names)
}
\arguments{
\item{names}{vector of data.frame column names}
}
\value{
a data.frame in which the Eikon formula's are replaced with the Eikon display Name which is the last part of the formula.
}
\description{
Convert eikon formula's in human readable names
}
