% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikonHelperfunctions.r
\name{GetISO103883_MIC}
\alias{GetISO103883_MIC}
\title{Download Operating mic defintions from the isowebsite
"https://www.iso20022.org/sites/default/files/ISO10383_MIC/ISO10383_MIC.csv"}
\usage{
GetISO103883_MIC()
}
\value{

}
\description{
Download Operating mic defintions from the isowebsite
"https://www.iso20022.org/sites/default/files/ISO10383_MIC/ISO10383_MIC.csv"
}
\examples{
GetISO103883_MIC()
}
