% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeCleaner.r
\name{make.true.NA_df}
\alias{make.true.NA_df}
\title{Remove <NA> NaN "" from data.table and replace with NA while retaining column class}
\usage{
make.true.NA_df(df)
}
\arguments{
\item{df}{data.frame with/without  <NA> NaN ""}
}
\value{
data.frame with only NA entries in the right column class
}
\description{
Remove <NA> NaN "" from data.table and replace with NA while retaining column class
}
\examples{
dfr<-data.frame(A=c(1,2,NaN,3),B=c("a", "<NA>" ,"","d"))
make.true.NA_df(dfr)
}
