% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{Process_RDP_output}
\alias{Process_RDP_output}
\title{Output processor for python function RDP$get_search_metadata}
\usage{
Process_RDP_output(python_json)
}
\arguments{
\item{metadata_python_df}{outcome of python RDP$get_search_metadata}
}
\value{
r data.frame
}
\description{
Output processor for python function RDP$get_search_metadata
}
\examples{
\dontrun{
"internal function don't use directly"
path = paste0("C:\\\\Users\\\\XXXX\\\\Documents\\\\GitHub",
"\\\\RefinitivR\\\\tests\\\\testthat\\\\PY_get_search_metadata_input.py")
PY_get_search_metadata_input <- reticulate::r_to_py(
 reticulate::py_load_object(file = path))
 r_df <- Process_RDP_output(PY_get_search_metadata_input)
}
}
\keyword{internal}
