% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_GetData.R
\name{rd_GetData}
\alias{rd_GetData}
\title{Function to obtain data from Eikon. Based on the Eikon python function get_data}
\usage{
rd_GetData(
  RDObject,
  rics,
  Eikonformulas,
  Parameters = NULL,
  raw_output = FALSE,
  time_out = 60,
  verbose = FALSE,
  SpaceConvertor = ".",
  use_field_names_in_headers = F
)
}
\arguments{
\item{RDObject}{Refinitiv Data connection object}

\item{rics}{a vector containing the instrument RICS}

\item{Eikonformulas}{a vector containing character string of Eikon Formulas}

\item{Parameters}{a named key value list for setting parameters, Default: NULL}

\item{raw_output}{to return the raw list by chunk for debugging purposes, default = FALSE}

\item{time_out}{set the maximum timeout to the Eikon server, default = 60}

\item{verbose}{boolean, set to true to print out the actual python call with time stamp for debugging.}

\item{SpaceConvertor}{converts spaces in variables name into one of the following characters ".", "," , "-", "_", default is "."}

\item{use_field_names_in_headers}{boolean return request fieldnames in stead of titles}
}
\value{
a data.frame containing data.from Eikon
}
\description{
The function automatically chunks the list of rics into chunks that comply with the api limitations and in the end rebuilds the chunks again into a single data.frame.
}
\details{
Currently there is a known bug in the reticulate package with large integers.
If a request is made that returns large integers the reticulate package will return -1 for these integers.
See also this issue \url{https://github.com/rstudio/reticulate/issues/323}
or try for yourself as indicated here \url{https://community.rstudio.com/t/large-integer-conversion-from-python-to-r/82568}
}
\references{
\url{https://developers.refinitiv.com/eikon-apis/eikon-data-api/docs?content=49692&type=documentation_item}
}
