% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{RDConnect}
\alias{RDConnect}
\alias{RDPConnect}
\title{RD connection function to refinitiv Data libraries}
\usage{
RDConnect(application_id = NA, PythonModule = "JSON", UUID = NA)

RDPConnect(application_id = NA, PythonModule = "JSON", UUID = NA)
}
\arguments{
\item{application_id}{refinitiv data api key}

\item{PythonModule}{character "JSON" or "RD"}

\item{UUID}{optional character parameter for custom instruments, not necessary for regular requests}
}
\value{
RD opbject
}
\description{
RD connection function to refinitiv Data libraries

RDPConnect alias to provide backwards compatability with code written for RDP
}
\examples{
\dontrun{
rd <- RDConnect(application_id = "your key")
}
\dontrun{
rd <- RDPConnect(application_id = "your key")
}
}
