% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinitiv.r
\name{RDConnect}
\alias{RDConnect}
\title{RD connection function to refinitiv Data libraries}
\usage{
RDConnect(application_id = NA, PythonModule = "RD")
}
\arguments{
\item{application_id}{refinitiv data api key}

\item{PythonModule}{character "JSON" or "RD"}
}
\value{
rdp opbject
}
\description{
RD connection function to refinitiv Data libraries
}
\examples{
\dontrun{
rd <- RDConnect(application_id = "your key")
}
}
