% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{RDPget_search_metadata}
\alias{RDPget_search_metadata}
\title{Get search metadata from RDP}
\usage{
RDPget_search_metadata(RDP = RDConnect(), searchView = NULL)
}
\arguments{
\item{RDP}{Refinitiv DataPlatform Connection object}

\item{searchView}{character choose from @seealso RDPShowAvailableSearchViews}
}
\value{
data.table with metadata search results
}
\description{
Get search metadata from RDP
}
\examples{
\dontrun{
test_rdp <- RDPget_search_metadata(RDP = RDPConnect(PythonModule = "RDP")
                              , searchView = "EquityQuotes")
test_json <- RDPget_search_metadata(RDP = RDPConnect(PythonModule = "JSON")
                              , searchView = "EquityQuotes")
test_rd <- RDPget_search_metadata(RDP = RDConnect()
                              , searchView = "EQUITY_QUOTES")
}
}
\seealso{
RDPShowAvailableSearchViews
}
