% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{RDPGetOptionAnalytics}
\alias{RDPGetOptionAnalytics}
\title{Get RDP option analytics}
\usage{
RDPGetOptionAnalytics(
  RDP = RDPConnect(),
  OptionRics = NULL,
  raw = FALSE,
  verbose = T
)
}
\arguments{
\item{RDP}{Refinitiv DataPlatform Connection object, defaults to  RDPConnect()}

\item{OptionRics}{character vector with option rics}

\item{raw}{return raw data from RDP call}

\item{verbose}{boolean, print download progress or not}
}
\value{
data.frame with option data
}
\description{
Get RDP option analytics
}
\examples{
\dontrun{
OPtionInstruments <- Refinitiv::RDPsearch(query = "aapl.o", view = "RelatedOption")
OPtionInstruments <- OPtionInstruments[grep(pattern = "*.U"
, x = OPtionInstruments$RIC),]$RIC

Analytics <- RDPGetOptionAnalytics(OptionRics = OPtionInstruments)
}
}
