% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpsearch.r
\name{RDPShowAvailableSearchViews}
\alias{RDPShowAvailableSearchViews}
\title{Show Available Search Views for seachview parameter in RDPget_search_metadata}
\usage{
RDPShowAvailableSearchViews(Platform = "JSON")
}
\arguments{
\item{Platform}{character vector either RD, RDP or JSON, defaults to JSON}
}
\value{
vector with searchviews that can be used.
}
\description{
Show Available Search Views for seachview parameter in RDPget_search_metadata
}
\examples{
RDPShowAvailableSearchViews(Platform = "JSON")
}
\seealso{
RDPget_search_metadata
}
