% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStreamConnection.r
\name{TestDataStreamCredentials}
\alias{TestDataStreamCredentials}
\title{Function to test if the DataStream Credentials are really working}
\usage{
TestDataStreamCredentials(DatastreamUsername, DatastreamPassword)
}
\arguments{
\item{DatastreamUsername}{character datastream Username}

\item{DatastreamPassword}{character datastream password}
}
\value{
TRUE if the DataStream credentials sare working and the JSON message when not.
}
\description{
Function to test if the DataStream Credentials are really working
}
\examples{
TestDataStreamCredentials(DatastreamUsername = "wrong username", DatastreamPassword = "wrong Password")
}
